backend/requirements.txt
fastapi
fastapi[standard]
google-genai
python-dotenv
uvicorn